+++ CHANGELOG.md
@@ -6,9 +6,15 @@ The format is loosely based on [Keep a Changelog](http://keepachangelog.com/).
 ## [v0.3] - Unreleased
 ### Added
 - Added QT-based GUI with controls for user input.
+- Added initial support for extended animation options. At the moment, it only allows you to set the threshold of lit LEDS in the Sparkle animation.
+- Added Utility class providing rand() and abs() (replaces stdlib).
+- Added Colors::MixMode::NONE. This is currently used to disable an Overlay and/or flag it as uninitialized.
 
 ### Removed
 - Removed GTK-based GUI
+
+### Known Issues
+- Segfault when switching the active Section to the Overlay. May need to add the Overlay as a new SectionController rather than just the Section.
 
 ## [v0.2] - 2017-05-01
 ### Added